feat: add extension for throw error
