+++ Runtime/Utils/ResponseExtension.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Newtonsoft.Json.Linq;
 using ReadyPlayerMe.Core;
 
 namespace ReadyPlayerMe.AvatarCreator
@@ -9,6 +10,19 @@ namespace ReadyPlayerMe.AvatarCreator
         {
             if (!response.IsSuccess)
             {
+                throw new Exception(response.Error);
+            }
+        }
+        
+        public static void ThrowIfError(this Response response)
+        {
+            if (!response.IsSuccess)
+            {
+                if (!string.IsNullOrEmpty(response.Text))
+                {
+                    var json = JObject.Parse(response.Text);
+                    throw new Exception(json["message"]!.ToString());
+                }
                 throw new Exception(response.Error);
             }
         }
